requirements.txt
asgiref==3.5.0
asttokens==2.0.5
backcall==0.2.0
cls-client==1.4.0
decorator==5.1.1
Django==4.0.1
django-livesync==0.5
django-tinymce==3.4.0
executing==0.8.2
ipython==8.0.1
jedi==0.18.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
parso==0.8.3
pathlib==1.0.1
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.0
platformdirs==2.5.1
prompt-toolkit==3.0.28
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.11.2
python-decouple==3.5
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
toml==0.10.2
tomli==2.0.1
tornado==6.1
traitlets==5.1.1
typing_extensions==4.1.1
watchdog==2.1.6
wcwidth==0.2.5
websocket-client==1.2.3
django-widget-tweaks==1.4.12